dev: refactored directories. added multi-admin feature in properties and application runner bean
