+++ Backend/src/main/java/com/praisesystem/backend/configuration/configs/security/jwt/JwtTokenFilter.java
@@ -1,4 +1,4 @@
-package com.praisesystem.backend.security.jwt;
+package com.praisesystem.backend.configuration.configs.security.jwt;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
