+++ Backend/src/main/java/com/praisesystem/backend/configuration/configs/security/jwt/JwtUser.java
@@ -1,4 +1,4 @@
-package com.praisesystem.backend.security.jwt;
+package com.praisesystem.backend.configuration.configs.security.jwt;
 
 import lombok.AccessLevel;
 import lombok.Value;
