+++ Backend/src/main/java/com/praisesystem/backend/configuration/properties/ApplicationProperties.java
@@ -5,11 +5,15 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
 import org.springframework.validation.annotation.Validated;
 
+import java.util.List;
+
 @Data
 @Validated
 @Component
 @ConfigurationProperties(prefix = "praise-system")
 public class ApplicationProperties {
 
-    String adminPublicKey;
+    List<String> adminAddresses;
+
+    String frontendUrl;
 }
