+++ Backend/src/main/java/com/praisesystem/backend/users/impl/UserTransactionalServiceImpl.java
@@ -2,12 +2,12 @@ package com.praisesystem.backend.users.impl;
 
 import com.praisesystem.backend.common.exceptions.exceptiontypes.NotFoundObjectException;
 import com.praisesystem.backend.configuration.properties.ApplicationProperties;
-import com.praisesystem.backend.roles.RoleService;
-import com.praisesystem.backend.roles.enums.RoleCode;
-import com.praisesystem.backend.roles.model.RoleEntity;
+import com.praisesystem.backend.users.roles.RoleService;
+import com.praisesystem.backend.users.roles.enums.RoleCode;
+import com.praisesystem.backend.users.roles.model.RoleEntity;
+import com.praisesystem.backend.users.UserRepository;
 import com.praisesystem.backend.users.dto.UserDto;
 import com.praisesystem.backend.users.mapper.UserMapper;
-import com.praisesystem.backend.users.UserRepository;
 import com.praisesystem.backend.users.model.UserEntity;
 import com.praisesystem.backend.users.services.UserTransactionalService;
 import lombok.AllArgsConstructor;
@@ -32,13 +32,24 @@ public class UserTransactionalServiceImpl implements UserTransactionalService {
     UserMapper userMapper;
 
     @Override
-    public UserEntity createAdmin() {
-        return userRepository.findByPublicKey(properties.getAdminPublicKey()).orElseGet(() -> {
-            List<RoleEntity> roles = roleService.findAll();
-            UserEntity newAdminUser = userMapper.toNewUserFromPublicKeyAndRoles(properties.getAdminPublicKey(), roles);
-            return userRepository.save(newAdminUser);
-        });
+    public void createAdmins() {
+        List<RoleEntity> roles = roleService.findAll();
+        List<String> adminAddresses = properties.getAdminAddresses();
 
+        List<String> existingAdmins = userRepository.findByPublicKeyIn(adminAddresses)
+                .stream()
+                .map(UserEntity::getPublicKey)
+                .collect(Collectors.toList());
+
+        List<UserEntity> newAdmins = adminAddresses.stream()
+                .distinct()
+                .filter(address -> !existingAdmins.contains(address))
+                .map(address -> userMapper.toNewUserFromPublicKeyAndRoles(address, roles))
+                .collect(Collectors.toList());
+
+        userRepository.saveAll(newAdmins)
+                .forEach(user -> log.info("[APPLICATION RUNNER] Admin with address ({}) successfully created.", user.getPublicKey()));
+        ;
     }
 
     @Override
