+++ Backend/src/main/java/com/praisesystem/backend/users/roles/mapper/RoleMapper.java
@@ -1,8 +1,8 @@
-package com.praisesystem.backend.roles.mapper;
+package com.praisesystem.backend.users.roles.mapper;
 
-import com.praisesystem.backend.roles.mapper.support.RoleToString;
-import com.praisesystem.backend.roles.mapper.support.RolesToStringList;
-import com.praisesystem.backend.roles.model.RoleEntity;
+import com.praisesystem.backend.users.roles.mapper.support.RoleToString;
+import com.praisesystem.backend.users.roles.mapper.support.RolesToStringList;
+import com.praisesystem.backend.users.roles.model.RoleEntity;
 import org.mapstruct.InjectionStrategy;
 import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
